Get rid of ArchiveModuleRegistry
