+++ spring-xd-dirt/src/test/resources/ResourceModuleRegistryLogicTests-modules/processor/dummy/README
