+++ spring-xd-dirt/src/test/resources/ResourceModuleRegistryLogicTests-modules/source/i-am-a-valid-module.jar/README
